requirements.txt
dbt==0.19.0
apache-airflow-providers-microsoft-azure==3.7.0
apache-airflow-providers-snowflake==2.5.0
riotwatcher
pandas